refactor

+++ src/ffi/converter.rs
@@ -37,10 +37,10 @@ pub(crate) unsafe fn to_totp(config: crate::ffi::totp_c_binds::TotpConfig) -> TO
 
     match TOTP::new(
         Algorithm::from_string(to_string(config.algorithm)),
-        Secret::new(to_str(config.secret)).unwrap(),      // TODO
-        NonZeroU8::new(config.length as u8).unwrap(),     // TODO
-        Radix::new(config.radix as u8).unwrap(),          // TODO
-        NonZeroU64::new(config.interval as u64).unwrap(), // TODO
+        Secret::new(to_str(config.secret)).unwrap(),  // TODO
+        NonZeroU8::new(config.length as u8).unwrap(), // TODO
+        Radix::new(config.radix as u8).unwrap(),      // TODO
+        NonZeroU64::new(config.interval).unwrap(),    // TODO
     ) {
         Ok(totp) => totp,
         Err(e) => panic!("{}", e),
